added route functionality to get save members to the database
