+++ controllers/api/userRoutes.js
@@ -1,5 +1,21 @@
 const router = require('express').Router();
-// const { User } = require('../../models');
+const { Member } = require('../../models');
+
+
+//Returns a list of all user. FOr development only
+router.get('/allUsers', async (req, res)=>{
+    try{
+        const allMembers = await Member.findAll();
+        res.status(200).json(allMembers); 
+    } catch(err) {
+        console.log(err);
+        res.status(500).json(err);
+    }
+});
+
+router.get('/login', (req, res)=>{
+    res.render('login');
+});
 
 router.post('/', async (req, res) => {
