+++ controllers/homeRoutes.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { Project, User } = require('../models');
+const { Project, Member } = require('../models');
 const withAuth = require('../utils/auth')
 
 
@@ -21,8 +21,21 @@ router.get('/signup', (req, res) => {
   res.render('signup')
 });
 
-router.post('/signup', (req, res) =>{
-    console.log(req.body);
+router.post('/signup', async (req, res) =>{
+  console.log(req.body);
+  try{
+    const newMember = await Member.create({
+    name: req.body.name,
+    username: req.body.username,
+    email: req.body.email,
+    password: req.body.password,
+    isProjectLead: false
+  });
+  res.status(200).json({member: newMember, message: 'New Member created'});
+  } catch (err){
+    console.log(err);
+    res.status(500).json(err);
+  }
 });
 
 module.exports = router;
