+++ seeds/projectLeadData.json
@@ -1,18 +1,18 @@
-[
-    {
-      "name": "Brandyn",
-      "email": "Brandyn@brandyn.com",
-      "password": "password12345"
-    },
-    {
-      "name": "Neil",
-      "email": "Neil@neil.com",
-      "password": "password12345"
-    },
-    {
-      "name": "Griselda",
-      "email": "Griselda@griselda.com",
-      "password": "password12345"
-    }
-  ]
+// [
+//     {
+//       "name": "Brandyn",
+//       "email": "Brandyn@brandyn.com",
+//       "password": "password12345"
+//     },
+//     {
+//       "name": "Neil",
+//       "email": "Neil@neil.com",
+//       "password": "password12345"
+//     },
+//     {
+//       "name": "Griselda",
+//       "email": "Griselda@griselda.com",
+//       "password": "password12345"
+//     }
+//   ]
