+++ views/layouts/main.handlebars
@@ -24,9 +24,9 @@
         <a href="/login">login</a>
         {{/if}} --}}
         <ul class="navbar-start">
-          <li><a class="navbar-item" href="">Login</a></li>
-          <li><a class="navbar-item" href="">Home</a></li>
-          <li><a class="navbar-item" href="">Sign Up</a></li>
+          <li><a class="navbar-item" href="/">Home</a></li>
+          <li><a class="navbar-item" href="/api/users/login">Login</a></li>
+          <li><a class="navbar-item" href="/signup">Sign Up</a></li>
         </ul>
       </nav>
     </header>
