+++ views/signup.handlebars
@@ -70,4 +70,4 @@
     </div>
 </form>
 
-<script src="./js/login.js"></script>
+<script src="./js/signup.js"></script>
